+++ src/data/stores/StoreD2ApiRepository.ts
@@ -1,0 +1,5 @@
+import { StoreRepository } from "../../domain/stores/repositories/StoreRepository";
+
+export class StoreD2ApiRepository implements StoreRepository {
+
+}
