+++ src/domain/stores/entities/Store.ts
