+++ src/domain/stores/entities/StorePermissions.ts
