+++ src/domain/stores/repositories/StoreRepository.ts
@@ -1,0 +1,3 @@
+export interface StoreRepository {
+
+}
