+++ src/domain/stores/usecases/DeleteStoreUseCase.ts
