+++ src/domain/stores/usecases/GetStoreUseCase.ts
