+++ src/domain/stores/usecases/ListStoresUseCase.ts
