+++ src/domain/stores/usecases/SetStoreAsDefaultUseCase.ts
