+++ src/domain/stores/usecases/ValidateStoreUseCase.ts
@@ -1,6 +1,6 @@
 import { UseCase } from "../../common/entities/UseCase";
 import { Store } from "../entities/Store";
-import { GitHubRepository } from "../repositories/GitHubRepository";
+import { GitHubRepository } from "../../packages/repositories/GitHubRepository";
 
 export class ValidateStoreUseCase implements UseCase {
     constructor(private githubRepository: GitHubRepository) {}
