+++ src/presentation/CompositionRoot.ts
@@ -5,7 +5,7 @@ import { MetadataD2ApiRepository } from "../data/metadata/MetadataD2ApiRepositor
 import { MetadataJSONRepository } from "../data/metadata/MetadataJSONRepository";
 import { GitHubOctokitRepository } from "../data/packages/GitHubOctokitRepository";
 import { DownloadWebRepository } from "../data/storage/DownloadWebRepository";
-import { StorageDataStoreRepository } from "../data/storage/StorageDataStoreRepository";
+import { StorageDataStoreRepository } from "../data/storage/StorageDataStoreClient";
 import { TransformationD2ApiRepository } from "../data/transformations/TransformationD2ApiRepository";
 import { AggregatedSyncUseCase } from "../domain/aggregated/usecases/AggregatedSyncUseCase";
 import { UseCase } from "../domain/common/entities/UseCase";
@@ -50,20 +50,20 @@ import { ListImportedPackagesUseCase } from "../domain/package-import/usecases/L
 import { SaveImportedPackagesUseCase } from "../domain/package-import/usecases/SaveImportedPackagesUseCase";
 import { CreatePackageUseCase } from "../domain/packages/usecases/CreatePackageUseCase";
 import { DeletePackageUseCase } from "../domain/packages/usecases/DeletePackageUseCase";
-import { DeleteStoreUseCase } from "../domain/packages/usecases/DeleteStoreUseCase";
+import { DeleteStoreUseCase } from "../domain/stores/usecases/DeleteStoreUseCase";
 import { DiffPackageUseCase } from "../domain/packages/usecases/DiffPackageUseCase";
 import { DownloadPackageUseCase } from "../domain/packages/usecases/DownloadPackageUseCase";
 import { GetPackageUseCase } from "../domain/packages/usecases/GetPackageUseCase";
 import { GetStorePackageUseCase } from "../domain/packages/usecases/GetStorePackageUseCase";
-import { GetStoreUseCase } from "../domain/packages/usecases/GetStoreUseCase";
+import { GetStoreUseCase } from "../domain/stores/usecases/GetStoreUseCase";
 import { ImportPackageUseCase } from "../domain/packages/usecases/ImportPackageUseCase";
 import { ListPackagesUseCase } from "../domain/packages/usecases/ListPackagesUseCase";
 import { ListStorePackagesUseCase } from "../domain/packages/usecases/ListStorePackagesUseCase";
-import { ListStoresUseCase } from "../domain/packages/usecases/ListStoresUseCase";
+import { ListStoresUseCase } from "../domain/stores/usecases/ListStoresUseCase";
 import { PublishStorePackageUseCase } from "../domain/packages/usecases/PublishStorePackageUseCase";
-import { SaveStoreUseCase } from "../domain/packages/usecases/SaveStoreUseCase";
-import { SetStoreAsDefaultUseCase } from "../domain/packages/usecases/SetStoreAsDefaultUseCase";
-import { ValidateStoreUseCase } from "../domain/packages/usecases/ValidateStoreUseCase";
+import { SaveStoreUseCase } from "../domain/stores/usecases/SaveStoreUseCase";
+import { SetStoreAsDefaultUseCase } from "../domain/stores/usecases/SetStoreAsDefaultUseCase";
+import { ValidateStoreUseCase } from "../domain/stores/usecases/ValidateStoreUseCase";
 import { Repositories } from "../domain/Repositories";
 import { DownloadFileUseCase } from "../domain/storage/usecases/DownloadFileUseCase";
 import { CreatePullRequestUseCase } from "../domain/synchronization/usecases/CreatePullRequestUseCase";
