+++ src/presentation/react/components/package-import-wizard/steps/InstanceStoreSelectionStep.tsx
@@ -1,7 +1,7 @@
 import { Box, Icon, IconButton } from "@material-ui/core";
 import React, { useState } from "react";
 import { PackageSource } from "../../../../../domain/package-import/entities/PackageSource";
-import { Store } from "../../../../../domain/packages/entities/Store";
+import { Store } from "../../../../../domain/stores/entities/Store";
 import i18n from "../../../../../locales";
 import {
     InstanceSelectionDropdown,
