+++ src/presentation/webapp/pages/store-creation/StoreCreationPage.tsx
@@ -10,7 +10,7 @@ import React, { useCallback, useEffect, useMemo, useState } from "react";
 import Linkify from "react-linkify";
 import { useHistory, useParams } from "react-router-dom";
 import { GitHubError } from "../../../../domain/packages/entities/Errors";
-import { Store } from "../../../../domain/packages/entities/Store";
+import { Store } from "../../../../domain/stores/entities/Store";
 import i18n from "../../../../locales";
 import { useAppContext } from "../../../react/contexts/AppContext";
 import PageHeader from "../../../react/components/page-header/PageHeader";
